Update in Lecture 2/15

+++ Lab1/src/pkgCore/Die.java
@@ -9,7 +9,7 @@ public class Die {
 		DieValue = (byte)(Math.random()*6 + 1);			
 	}
 
-	public int getDieValue() {
+	public byte getDieValue() {
 		return DieValue;
 	}
 }
